src/risk-assessment/services/real-time-monitoring.service.ts
import { Injectable, Logger } from '@nestjs/common';
import { CustomerDto } from '../dto/customer.dto';
import { RiskLevel } from '../enums/risk-level.enum';
import { AuditLogService } from './audit-log.service';

type ActionCategory =
  | 'RISK_ASSESSMENT'
  | 'DOCUMENTATION'
  | 'VERIFICATION'
  | 'CONFIGURATION';

interface MonitoringAlert {
  customerId: number;
  alertType: string;
  severity: 'LOW' | 'MEDIUM' | 'HIGH';
  description: string;
  timestamp: Date;
}

@Injectable()
export class RealTimeMonitoringService {
  private readonly logger = new Logger(RealTimeMonitoringService.name);
  private readonly alertThresholds = {
    transactionAmount: 100000,
    highRiskTransactionsPerDay: 3,
    suspiciousActivityWindow: 24, // hours
  };

  constructor(private readonly auditLogService: AuditLogService) {}

  async monitorCustomerActivity(
    customer: CustomerDto,
    activityType: string,
    activityData: any,
  ): Promise<MonitoringAlert[]> {
    try {
      const alerts: MonitoringAlert[] = [];

      // Check for suspicious patterns
      const suspiciousPatterns = await this.detectSuspiciousPatterns(
        customer.id,
        activityType,
        activityData,
      );

      if (suspiciousPatterns.length > 0) {
        alerts.push(...suspiciousPatterns);
        await this.handleAlerts(customer, alerts);
      }

      // Risk level escalation check
      if (customer.riskLevel !== RiskLevel.HIGH) {
        const shouldEscalate = await this.checkRiskEscalation(
          customer.id,
          alerts,
        );

        if (shouldEscalate) {
          await this.escalateRiskLevel(customer);
        }
      }

      return alerts;
    } catch (error) {
      this.logger.error(
        `Error monitoring customer ${customer.id} activity`,
        error.stack,
      );
      throw error;
    }
  }

  private async detectSuspiciousPatterns(
    customerId: number,
    activityType: string,
    activityData: any,
  ): Promise<MonitoringAlert[]> {
    const alerts: MonitoringAlert[] = [];
    const timestamp = new Date();

    // Implement pattern detection logic here
    // This is a simplified example
    if (activityData.amount > this.alertThresholds.transactionAmount) {
      alerts.push({
        customerId,
        alertType: 'LARGE_TRANSACTION',
        severity: 'MEDIUM',
        description: 'Unusually large transaction detected',
        timestamp,
      });
    }

    return alerts;
  }

  private async handleAlerts(
    customer: CustomerDto,
    alerts: MonitoringAlert[],
  ): Promise<void> {
    for (const alert of alerts) {
      await this.auditLogService.logAction(
        customer.id.toString(),
        'ALERT_GENERATED',
        'RISK_ASSESSMENT',
        'SYSTEM',
        alert.description,
      );
    }
  }

  private async checkRiskEscalation(
    customerId: number,
    recentAlerts: MonitoringAlert[],
  ): Promise<boolean> {
    const highSeverityAlerts = recentAlerts.filter(
      (alert) => alert.severity === 'HIGH',
    );

    return (
      highSeverityAlerts.length >=
      this.alertThresholds.highRiskTransactionsPerDay
    );
  }

  private async escalateRiskLevel(customer: CustomerDto): Promise<void> {
    await this.auditLogService.logAction(
      customer.id.toString(),
      'RISK_ESCALATION',
      'RISK_ASSESSMENT',
      'SYSTEM',
      'Risk level escalated due to suspicious activity',
    );
  }
}
